+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2026-01-07
+
 ### Added
 
 - Added `JwsSigner::key_id`.
@@ -46,6 +48,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Binary, string, base64, hex secret encodings.
 - Auto-selection of send/sync based on platform.
 
-[unreleased]: https://github.com/chewie-rs/chewie-crypto/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/chewie-rs/chewie-crypto/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/chewie-rs/chewie-crypto/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/chewie-rs/chewie-crypto/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/chewie-rs/chewie-crypto/tag/v0.1.0